backend-ml/requirements.txt
flask==3.1.0
flask-cors==5.0.0
pandas==2.2.3
numpy==2.2.2
yfinance==0.2.54
requests==2.32.3
scikit-learn==1.6.1
python-dotenv==1.0.1
